Connecting w/Segment; export of interaction data
