+++ src/pyproject.toml
@@ -79,6 +79,8 @@ pdf2doi = "1.5"
 beautifulsoup4 = "4.12.2"
 django-allauth = "0.48.0"
 bibtexparser = {version = "^2.0.0b3", allow-prereleases = true}
+pyyaml = "6.0.1"
+segment-analytics-python = "^2.2.3"
 
 [tool.poetry.dev-dependencies]
