+++ src/user/related_models/action_model.py
@@ -6,6 +6,7 @@ from django.db.models import DecimalField, Sum
 from django.db.models.functions import Coalesce
 from django.utils import timezone
 
+from analytics.utils.analytics_mapping_utils import log_analytics_event
 from discussion.models import Comment, Reply, Thread
 from hub.models import Hub
 from paper.models import Paper, PaperSubmission
@@ -54,16 +55,19 @@ class Action(DefaultModel):
         )
 
     def save(self, *args, **kwargs):
-        if self.id is None:
-            from mailing_list.tasks import notify_immediate
+        from mailing_list.tasks import notify_immediate
 
-            super().save(*args, **kwargs)
+        is_new = self.id is None
+        super().save(*args, **kwargs)
+
+        if is_new:
+            print("newly created", self)
+            log_analytics_event(self)
+
             if not TESTING:
                 notify_immediate.apply_async((self.id,), priority=5)
             else:
                 notify_immediate(self.id)
-        else:
-            super().save(*args, **kwargs)
 
     def set_read(self):
         self.read_date = timezone.now()
